+++ tests/dir-surround/lua/nvim-treesitter/parsers.lua
@@ -1,0 +1,11 @@
+local parser = {
+  for_each_tree = function(_, f)
+    local tree = { root = function(_) return {} end }
+    local lang_tree = { lang = function(_) return 'lua' end }
+    f(tree, lang_tree)
+  end,
+}
+
+local get_parser = function(_) return parser end
+
+return { get_parser = get_parser }